requirements.txt
streamlit==1.37.1
pillow==10.4.0
sqlalchemy==2.0.34
python-dotenv==1.0.1
google-genai==0.2.0